Сделал установку времени тех работ по таймеру и с админ панели.
